First Year Project
"Electromania" is an innovative ecommerce project focused on providing a convenient platform for customers to explore and purchase a diverse range of electronics, including mobile phones, laptops, and home supplies. The website has been meticulously designed using a combination of HTML, CSS, JavaScript and some Bootstrap ensuring an attractive and user-friendly interface. The project's backend functionality is powered by PHP, guaranteeing a secure and efficient checkout process.
A notable feature of the project is the strategic integration of PHPMailer. This integration facilitates the automated delivery of emails exclusively during the checkout process.